+++ servertools/common/misc_functions.py
@@ -28,3 +28,21 @@ def version():
     """ Shows the version of the application on the terminal """
     application_version = pkg_resources.require("servertools")[0].version
     print(f"You are running version {application_version} ")
+
+
+def show_port_open_or_close(self, result: int, port: int, silence: bool = False):
+    """
+        Depending on the response parsed to the function prints if the
+        port is open or closed, the closed ones can be silenced with the
+        True flag.
+
+        Arguments:
+            result {int}: The result of the socket function
+            port {int}: The scanned port
+            silence {bool}: The flag to make close ports not to show
+    """
+    if result == 0:
+        print(f"Port {port}: 	 Open")
+    else:
+        if not silence:
+            print(f"Port {port}: 	 Close")
