+++ servertools/variables/logos.py
